4 templates
removed old activator listener template.
Added new version of Activator listener. This one uses the API to provide metadata.
Added Notification Center template.
Added Flipswitch template. It is already set up to apply and reflect changes based on the preferences file saved in NSHomeDirectory()/Library/Preferences/ and used by the settings app if the tweak uses PreferenceLoader to add a cell inside it.
Added Tweak template. Modified version of the one packaged with Theos, adds built in support for preferences, using a single switch in the root level of settings.